test(ui): design-token smoke test + lint/verify pass
Add index.theme.test.ts (light/dark applyTheme wiring).
Fix react-refresh/only-export-components lint error in GenreIcon.tsx by
extracting canonicalizeGenre + PATTERNS into genreUtils.ts; update the
test import accordingly. All 80 tests pass, build clean, lint clean.
